+++ lite/kernels/mlu/bridges/act_op_test.cc
@@ -44,40 +44,40 @@ void act_ref(const std::shared_ptr<operators::ActivationOp> op) {
 
   // "sigmoid","relu","tanh","relu_clipped","leaky_relu","softsign","hard_sigmoid"
   if (op_type == "sigmoid") {
-    for (size_t i = 0; i < out->numel(); i++) {
+    for (int i = 0; i < out->numel(); i++) {
       out_data[i] = 1.f / (1.f + std::exp(-x_data[i]));
     }
   } else if (op_type == "relu") {
-    for (size_t i = 0; i < out->numel(); i++) {
+    for (int i = 0; i < out->numel(); i++) {
       out_data[i] = std::max(0.f, x_data[i]);
     }
   } else if (op_type == "tanh") {
-    for (size_t i = 0; i < out->numel(); i++) {
+    for (int i = 0; i < out->numel(); i++) {
       out_data[i] = (std::exp(x_data[i]) - std::exp(-x_data[i])) /
                     (std::exp(x_data[i]) + std::exp(-x_data[i]));
     }
   } else if (op_type == "relu_clipped") {
     auto relu_clipped_coef = op_info->GetAttr<float>("Relu_clipped_coef");
-    for (size_t i = 0; i < out->numel(); i++) {
+    for (int i = 0; i < out->numel(); i++) {
       out_data[i] = std::min(std::max(0.f, x_data[i]), relu_clipped_coef);
     }
   } else if (op_type == "relu6") {
-    for (size_t i = 0; i < out->numel(); i++) {
+    for (int i = 0; i < out->numel(); i++) {
       out_data[i] = std::min(std::max(0.f, x_data[i]), 6.f);
     }
   } else if (op_type == "leaky_relu") {
     auto alpha = op_info->GetAttr<float>("alpha");
-    for (size_t i = 0; i < out->numel(); i++) {
+    for (int i = 0; i < out->numel(); i++) {
       out_data[i] = std::max(x_data[i], x_data[i] * alpha);
     }
   } else if (op_type == "softsign") {
-    for (size_t i = 0; i < out->numel(); i++) {
+    for (int i = 0; i < out->numel(); i++) {
       out_data[i] = x_data[i] / (1 + std::abs(x_data[i]));
     }
   } else if (op_type == "hard_sigmoid") {
     auto slope = op_info->GetAttr<float>("slope");
     auto offset = op_info->GetAttr<float>("offset");
-    for (size_t i = 0; i < out->numel(); i++) {
+    for (int i = 0; i < out->numel(); i++) {
       out_data[i] = std::min(1.f, slope * x_data[i] + offset);
       out_data[i] = std::max(0.f, out_data[i]);
     }
